+++ docs/CHANGELOG.rst
@@ -14,6 +14,10 @@ Fixed
 -----
 * Use new import paths in ``resolwe.flow``
 
+Changed
+-------
+* Support Resolwe test framework
+
 
 ================
 1.3.0 2017-01-28